requirements.txt
patchright==1.58.0
tf-playwright-stealth==1.2.0
browser-use==0.12.6
mcp==1.26.0
HumanCursor==1.1.5
python-dotenv==1.2.1
